requirements.txt
# 仅保留核心依赖，确保稳定运行
requests==2.32.3          # 稳定版本：链接请求核心依赖
python-dotenv==1.0.1      # 稳定版本：读取.env配置文件